Updated playing card cover
