--- next-app/src/components/Playing-Cards-Page/PlayingCards.js
@@ -1,125 +1,0 @@
-'use client';
-
-import { useEffect, useRef } from 'react';
-
-import ReactLenis from '@studio-freight/react-lenis';
-import gsap from 'gsap';
-import ScrollTrigger from 'gsap/ScrollTrigger';
-import { useGSAP } from '@gsap/react';
-
-import Card from './Card';
-import './playing-cards.css';
-
-gsap.registerPlugin(ScrollTrigger);
-
-const PlayingCards = () => {
-  const container = useRef(null);
-  const cardRefs = useRef([]);
-
-  // Card spread and rotate animation that plays on-scroll
-  useGSAP(
-    () => {
-      const cards = cardRefs.current;
-      const totalScrollHeight = window.innerHeight * 3;
-      const positions = [14, 38, 62, 86]; // card position on-screen
-      const rotations = [-15, -7.5, 7.5, 15]; // card rotation offset on-screen
-
-      // pin cards section
-      ScrollTrigger.create({
-        trigger: container.current.querySelector('.cards'),
-        start: 'top top',
-        end: () => `+=${totalScrollHeight}`,
-        pin: true,
-        pinSpacing: true,
-      });
-
-      // spread cards
-      cards.forEach(
-        (card, index) => {
-          gsap.to(card, {
-            left: `${positions[index]}%`,
-            rotation: `${rotations[index]}%`,
-            ease: 'none',
-            scrollTrigger: {
-              trigger: container.current.querySelector('.cards'),
-              start: 'top top',
-              end: () => `+=${window.innerHeight}`,
-              scrub: 0.5,
-              id: `spread-${index}`,
-            },
-          });
-        },
-        { scope: container }
-      );
-
-      // flip cards and reset rotations with stagger
-      cards.forEach((card, index) => {
-        const frontEl = card.querySelector('.flip-card-front');
-        const backEl = card.querySelector('.flip-card-back');
-
-        const staggerOffset = index * 0.05;
-
-        const startOffset = 1 / 3 + staggerOffset;
-        const endOffset = 2 / 3 + staggerOffset;
-
-        ScrollTrigger.create({
-          trigger: container.current.querySelector('.cards'),
-          start: 'top top',
-          end: () => `+=${totalScrollHeight}`,
-          scrub: 1,
-          id: `rotate-flip-${index}`,
-          onUpdate: (self) => {
-            const progress = self.progress;
-
-            if (progress >= startOffset && progress <= endOffset) {
-              const animationProgress = (progress - startOffset) / (1 / 3);
-              const frontRotation = -180 * animationProgress;
-              const backRotation = 180 - 180 * animationProgress;
-              const cardRotation = rotations[index] * (1 - animationProgress);
-
-              gsap.to(frontEl, { rotateY: frontRotation, ease: 'power1.out' });
-              gsap.to(backEl, { rotateY: backRotation, ease: 'power1.out' });
-              gsap.to(card, {
-                xPercent: -50,
-                yPercent: -50,
-                rotate: cardRotation,
-                ease: 'power1.out',
-              });
-            }
-          },
-        });
-      });
-    },
-    { scope: container }
-  );
-
-  // Cleanup function on component unmount
-  useEffect(() => {
-    return () => {
-      ScrollTrigger.getAll().forEach((trigger) => trigger.kill());
-    };
-  }, []);
-
-  return (
-    <>
-      <ReactLenis root>
-        <div className="container" ref={container}>
-          <section className="cards">
-            {[...Array(4)].map((_, index) => (
-              <Card
-                key={index}
-                id={`card-${index + 1}`}
-                frontSrc="/images/playing-card-cover.png"
-                frontAlt="Card Image"
-                backText="Your card details appear here"
-                ref={(el) => (cardRefs.current[index] = el)}
-              />
-            ))}
-          </section>
-        </div>
-      </ReactLenis>
-    </>
-  );
-};
-
-export default PlayingCards;
