+++ next-app/src/components/Playing-Cards-Page/playing-cards.css
@@ -13,6 +13,8 @@ html, body {
 }
 
 .container {
+    position: relative;
+    padding-top: 200px;
     width: 100%;
     height: 100%;
 }
